add fixture proxy_arp_enabled
